move some files into a security namespace/folder
