+++ DDDEastAnglia/Controllers/ResetPasswordController.cs
@@ -3,6 +3,7 @@ using System.Web.Mvc;
 using DDDEastAnglia.DataAccess;
 using DDDEastAnglia.Helpers;
 using DDDEastAnglia.Helpers.Email;
+using DDDEastAnglia.Helpers.Security;
 using DDDEastAnglia.Models;
 
 namespace DDDEastAnglia.Controllers
