+++ DDDEastAnglia/DDDEastAnglia.csproj
@@ -397,7 +397,7 @@
     <Compile Include="Domain\IConference.cs" />
     <Compile Include="Helpers\AlphabeticalSort.cs" />
     <Compile Include="Helpers\Context\HttpContextControllerInformationProvider.cs" />
-    <Compile Include="Helpers\IResetPasswordThingy.cs" />
+    <Compile Include="Helpers\Security\IResetPasswordThingy.cs" />
     <Compile Include="Helpers\ISessionSorter.cs" />
     <Compile Include="Helpers\ISortAlgorithm.cs" />
     <Compile Include="Helpers\ListExtensions.cs" />
@@ -419,7 +419,7 @@
     <Compile Include="Helpers\RandomSort.cs" />
     <Compile Include="Helpers\SessionSorter.cs" />
     <Compile Include="Helpers\Email\Smtp\SmtpHostSettings.cs" />
-    <Compile Include="Helpers\WebSecurityWrapper.cs" />
+    <Compile Include="Helpers\Security\WebSecurityWrapper.cs" />
     <Compile Include="Helpers\MvcHelperModule.cs" />
     <Compile Include="Models\BannerViewModel.cs" />
     <Compile Include="Models\BannerModel.cs" />
